cfn-template-manager-mcp-server/pyproject.toml
[project]
name = "awslabs-cfn-template-manager-mcp-server"
version = "0.1.0"
description = "MCP server for managing AWS CloudFormation deployments with pre-existing templates"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "AWS Labs" }
]
license = { text = "Apache-2.0" }
dependencies = [
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "fastmcp>=0.2.3",
    "loguru>=0.7.2",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
cfn-template-manager-mcp-server = "awslabs.cfn_template_manager.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["awslabs"]

